+++ example/packages/api/src/createServer.ts
@@ -4,8 +4,9 @@ import {
   type ServerResponse,
 } from "node:http";
 import { URL } from "node:url";
-import { toJSONSchema, z } from "zod/v4";
+import type { z } from "zod/v4";
 import { createConfig } from "./config.js";
+import { controller as healthController } from "./controllers/health/get.js";
 import { controller as getNoteController } from "./controllers/notes/[noteId]/get.js";
 import { controller as listNotesController } from "./controllers/notes/get.js";
 import { controller as createNoteController } from "./controllers/notes/post.js";
@@ -20,6 +21,7 @@ import type {
   ServerLifecycle,
 } from "./types.js";
 import { sendJson } from "./utils/http.js";
+import { createOpenApiController } from "./utils/openapi.js";
 
 const documentedControllers = [
   listNotesController,
@@ -28,61 +30,6 @@ const documentedControllers = [
 ] satisfies readonly ControllerDefinition<ControllerSchema>[];
 
 type ZodObjectSchema = z.ZodObject<Record<string, z.ZodTypeAny>>;
-
-function createHealthController(): ControllerDefinition<ControllerSchema> {
-  const schema: ControllerSchema = {
-    method: "GET",
-    path: "/health",
-    summary: "Readiness probe",
-    tags: ["Health"],
-    responses: {
-      200: {
-        description: "Service is healthy",
-        content: {
-          "application/json": {
-            schema: z.object({ status: z.literal("healthy") }),
-          },
-        },
-      },
-    },
-  };
-
-  return {
-    schema,
-    handler: async ({ response }) => {
-      sendJson(response, 200, { status: "healthy" });
-    },
-  };
-}
-
-function createOpenApiController(
-  controllers: readonly ControllerDefinition<ControllerSchema>[],
-): ControllerDefinition<ControllerSchema> {
-  const schema: ControllerSchema = {
-    method: "GET",
-    path: "/openapi",
-    summary: "OpenAPI document",
-    tags: ["Documentation"],
-    responses: {
-      200: {
-        description: "Generated OpenAPI document",
-        content: {
-          "application/json": {
-            schema: z.object({}).passthrough(),
-          },
-        },
-      },
-    },
-  };
-
-  return {
-    schema,
-    handler: async ({ context, response }) => {
-      const document = generateOpenApiDocument(context.config, controllers);
-      sendJson(response, 200, document);
-    },
-  };
-}
 
 function buildRoutes(
   controllers: readonly ControllerDefinition<ControllerSchema>[],
@@ -158,113 +105,6 @@ function parseQuery<TSchema extends ControllerSchema>(
   return result.data as InferQuery<TSchema>;
 }
 
-function buildOpenApiOperation(
-  controller: ControllerDefinition<ControllerSchema>,
-) {
-  const { schema } = controller;
-  const parameters: Array<Record<string, unknown>> = [];
-
-  if (schema.params) {
-    const entries = Object.entries(schema.params.shape) as Array<
-      [string, z.ZodTypeAny]
-    >;
-    for (const [name, definition] of entries) {
-      parameters.push({
-        name,
-        in: "path",
-        required: true,
-        schema: toJSONSchema(definition),
-      });
-    }
-  }
-
-  if (schema.query) {
-    const entries = Object.entries(schema.query.shape) as Array<
-      [string, z.ZodTypeAny & { isOptional?: () => boolean }]
-    >;
-    for (const [name, definition] of entries) {
-      const typed = definition;
-      parameters.push({
-        name,
-        in: "query",
-        required: !(
-          typeof typed.isOptional === "function" && typed.isOptional()
-        ),
-        schema: toJSONSchema(typed),
-      });
-    }
-  }
-
-  const responses = Object.fromEntries(
-    Object.entries(schema.responses).map(([status, response]) => {
-      const jsonContent = response.content?.["application/json"];
-      return [
-        status,
-        {
-          description: response.description,
-          ...(jsonContent
-            ? {
-                content: {
-                  "application/json": {
-                    schema: toJSONSchema(jsonContent.schema),
-                  },
-                },
-              }
-            : {}),
-        },
-      ];
-    }),
-  );
-
-  const requestBody = schema.body
-    ? {
-        required: true,
-        content: {
-          [schema.body.contentType]: {
-            schema: toJSONSchema(schema.body.schema as z.ZodTypeAny),
-          },
-        },
-      }
-    : undefined;
-
-  return {
-    summary: schema.summary,
-    description: schema.description,
-    tags: schema.tags,
-    ...(parameters.length > 0 ? { parameters } : {}),
-    ...(requestBody ? { requestBody } : {}),
-    responses,
-  };
-}
-
-function generateOpenApiDocument(
-  config: Config,
-  controllers: readonly ControllerDefinition<ControllerSchema>[],
-) {
-  const paths: Record<string, Record<string, unknown>> = {};
-
-  for (const controller of controllers) {
-    const method = controller.schema.method.toLowerCase();
-    let pathItem = paths[controller.schema.path];
-    if (!pathItem) {
-      pathItem = {};
-      paths[controller.schema.path] = pathItem;
-    }
-    pathItem[method] = buildOpenApiOperation(controller);
-  }
-
-  return {
-    openapi: "3.1.0",
-    info: {
-      version: "0.0.0",
-      title: "Notes API",
-      description: "Example implementation for the NodeJS Code Guide",
-    },
-    servers: [{ url: `http://${config.host}:${config.port}` }],
-    paths,
-  };
-}
-
 async function handleController<TSchema extends ControllerSchema>(
   controller: ControllerDefinition<TSchema>,
   context: AppContext,
@@ -288,7 +128,6 @@ export function createServer(options?: {
     options?.contextFactory ??
     ((configuration) => buildContext({ config: configuration }));
 
-  const healthController = createHealthController();
   const openApiController = createOpenApiController(documentedControllers);
 
   const routes: Route[] = buildRoutes([
