--- example/packages/api/src/globals.d.ts
@@ -1,11 +1,0 @@
-declare class URLPattern {
-  constructor(init?: { pathname?: string }, baseURL?: string);
-  exec(input?: string | URL, baseURL?: string): URLPatternResult | null;
-  test(input?: string | URL, baseURL?: string): boolean;
-}
-
-declare interface URLPatternResult {
-  pathname: {
-    groups: Record<string, string>;
-  };
-}
